pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "A Python-based web scraping API built with FastAPI that provides easy access to manga contents."
authors = ["tokitou-san <tokitousan2005@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.2"
fastapi = "^0.110.3"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
requests = "^2.31.0"
pydantic-core = "^2.18.2"
anyio = "^4.2.0"
black = "^24.4.2"
pre-commit = "^3.7.0"
jinja2 = "^3.1.3"
markdown = "^3.6"
selectolax = "^0.3.21"
httpx = "^0.27.0"

[tool.poetry.scripts]
dev = "app.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
